% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_elbow.R
\name{cluster.elbow}
\alias{cluster.elbow}
\title{Visually inspect the optimal numbers of clusters}
\usage{
cluster.elbow(
  dat,
  FUNcluster,
  method = c("wss", "silhouette", "gap_stat"),
  k.max = 10,
  ...
)
}
\arguments{
\item{dat}{an input data.}

\item{FUNcluster}{a partitioning function which accepts as first argument a (data) matrix like x, second argument, say k, k >= 2, the number of clusters desired, and returns a list with a component named cluster which contains the grouping of observations. By default, "kmeans".}

\item{method}{the method to be used for estimating the optimal number of clusters. Possible values are "silhouette" (for average silhouette width), "wss" (for total within sum of square) and "gap_stat" (for gap statistics).}

\item{k.max}{the maximum number of clusters to consider. By default, 10.}

\item{...}{optional arguments. See or directly use \code{factoextra::fviz_nbclust}.}
}
\value{
\code{cluster.elbow} returns a ggplot2 object.
}
\description{
A convenient wrapper for \code{factoextra::fviz_nbclust}, which plot within cluster sums of squares, average silhouette and gap statistics for clustering.
The required parameters and their corresponding documentations are simplified versions of that in  \code{factoextra::fviz_nbclust}.
}
\examples{
\dontrun{
data(coptm)
cluster.elbow(dat=optm, FUNcluster=kmeans, method="wss", k.max=10, linecolor="black")
}
}
\author{
Neo Christopher Chung \email{nchchung@gmail.com}
}
