% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cys_optm_missing}
\alias{cys_optm_missing}
\title{Cysteine oxidative post-translational modifications (O-PTM) of proteomes in cardiac hypertrophyhh, with missing values}
\format{A data frame with 848 rows (O-PTM sites) and 9 columns:
\describe{
  \item{column 1-3}{meta data for O-PTM sites}
  \item{column 4-9}{ratios of occupancies in ISO over that in control conditions, from 1 to 14 days}
  ...
}}
\source{
Peipei Ping lab at UCLA
}
\usage{
cys_optm_missing
}
\description{
The temporal changes of 848 Cysteine O-PTMs over 14 days. This data contains missing values.
It contains three different types of Cys modifications: reversible cysteine O-PTM, cysteine sulfinylation (CysSO2H), and cysteine sulfonylation (CysSO3H).
}
\keyword{datasets}
