% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise_pca.R
\name{denoise_pca}
\alias{denoise_pca}
\alias{denoise_svd}
\title{Deonise the input data by principal component analysis (PCA)}
\usage{
denoise_pca(
  dat,
  timepoints = NULL,
  r = NULL,
  center.dat = TRUE,
  scale.dat = FALSE,
  verbose = FALSE,
  seed = NULL,
  ...
)

denoise_svd(
  dat,
  timepoints = NULL,
  r = NULL,
  center.dat = TRUE,
  scale.dat = FALSE,
  verbose = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{dat}{a time-series data matrix with \code{m} biomarkers as rows, over \code{n} time points (columns).}

\item{timepoints}{a vector of time points for columns of dat.}

\item{r}{the number of PCs or eigenmatrices to retain.}

\item{center.dat}{a logical specifying to center the input and denoised data. By default, \code{TRUE}.}

\item{scale.dat}{a logical specifying to scale the input and denoised data. By default, \code{FALSE}.}

\item{verbose}{a logical specifying to print the computational progress. By default, \code{FALSE}.}

\item{seed}{a seed for the random number generator.}

\item{...}{optional arguments.}
}
\value{
\code{denoise_pca} returns a matrix of denoised data.
}
\description{
The data is denoised by using the top r eigenmatrices.
After apply SVD on the centered and/or scaled data, the top r eigenmatrices are constructed where is r < (n,m).
}
\details{
If the input data has missing values, the SVDImpute algorithm (Troyanskaya et al. 2001) is used.
}
\examples{
\dontrun{
data(cys_optm)
meta <- cys_optm[,1:3]
optm <- log(cys_optm[,4:9])
days <- as.numeric(colnames(optm))

denoised_optm <- denoise_pca(optm, timepoints = days, r=3)
}
}
\references{
Troyanskaya, O., Cantor, M., Sherlock, G., Brown, P., Hastie, T., Tibshirani, R., Botstein, D. and Altman, R.B. (2001). Missing value estimation methods for DNA microarrays. Bioinformatics 17(6), 520–525.
}
\author{
Neo Christopher Chung \email{nchchung@gmail.com}
}
