% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{dispersion}
\alias{dispersion}
\title{Calculate dispersion metrics}
\usage{
dispersion(
  dat,
  n_bins = 10,
  center.dat = TRUE,
  scale.dat = FALSE,
  verbose = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{dat}{a time-series data matrix with \code{m} biomarkers as rows, over \code{n} time points (columns).}

\item{center.dat}{a logical specifying to center the input and denoised data. By default, \code{TRUE}.}

\item{scale.dat}{a logical specifying to scale the input and denoised data. By default, \code{FALSE}.}

\item{verbose}{a logical specifying to print the computational progress. By default, \code{FALSE}.}

\item{seed}{a seed for the random number generator.}

\item{...}{optional arguments.}

\item{timepoints}{a vector of time points for columns of dat.}

\item{dof}{the degree of freedom in spline.smooth. By default, \code{cv} performs cross-validation per-variable.}
}
\value{
\code{dispersion} returns a data.frame of statistics for \code{m} variables (rows).
}
\description{
For each variable (row), calculate a dispersion (var/mean) and a z-score of normalized dispersion.
A z-score of normalized dispersion is motivated due to correlation between means and variances.
Variables are binned (by default,\code{n_bins=10}) by their means. Within a bin, a z-score is calculated.
It's recommended that you look at both results \code{zdisp} and \code{disp} to decide how to filter the data.
}
\examples{
\dontrun{
data(cys_optm)
meta <- cys_optm[,1:4]
optm <- log(cys_optm[meta$Select,5:10])
optm <- t(scale(t(optm), scale=TRUE, center=TRUE))
days <- as.numeric(colnames(optm))

disp_optm <- dispersion(optm, timepoints = days, dof="cv")
disp_optm <- cbind(meta, disp_optm)
# make a histogram of dispersion statistics
hist(disp_optm$disp, 100)
# make a histogram of z-score of normalized dispersion
hist(disp_optm$zdisp, 100)
# library(readr)
# write_excel_csv(disp_optm, file="~/coptm_dispersion.csv")
}
}
\author{
Neo Christopher Chung \email{nchchung@gmail.com}
}
