% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_elbow.R
\name{pca.elbow}
\alias{pca.elbow}
\title{Visualize the percent variance explained by PCs}
\usage{
pca.elbow(
  dat,
  center.dat = TRUE,
  scale.dat = FALSE,
  r = NULL,
  pt.size = 1,
  title = ""
)
}
\arguments{
\item{dat}{a data matrix with \code{m} biomarkers as rows, over \code{n} time points (columns).}

\item{center.dat}{a logical specifying to center the data. By default, \code{TRUE}.}

\item{scale.dat}{a logical specifying to scale the data. By default, \code{FALSE}.}

\item{r}{the top \code{r} PCs to display. by default, showing all PCs.}

\item{pt.size}{the size of data points.}

\item{title}{the title for the resulting plot.}
}
\value{
\code{pca.elbow} returns a ggplot2 object.
}
\description{
Plot the classic elbow plot, showing the percent variance explained by PCs.
}
\details{
The input data is either centered and/or scaled.
SVD/PCA is computed and the percent variacnes explained by the top \code{r} PCs are plotted.
If the denoising step is used, please input the denoised data.
}
\examples{
\dontrun{
data(cys_optm)
meta <- cys_optm[,1:4]
optm <- log(cys_optm[meta$Select,5:10])
optm <- t(scale(t(optm), scale=TRUE, center=TRUE))
days <- as.numeric(colnames(optm))

pca.elbow(dat=optm,
center.dat=TRUE,
scale.dat=FALSE,
title="Percent variance explained by 10 PCs")
}
}
\author{
Neo Christopher Chung \email{nchchung@gmail.com}
}
