% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_pca.R
\name{preprocess_pca}
\alias{preprocess_pca}
\title{Preprocess the input data by principal component analysis (PCA) and related methods}
\usage{
preprocess_pca(
  dat,
  r = NULL,
  method = c("em", "nipals"),
  center.dat = TRUE,
  scale.dat = FALSE,
  verbose = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{dat}{a time-series data matrix with \code{m} biomarkers as rows, over \code{n} time points (columns).}

\item{r}{the number of PCs or eigenmatrices to retain.}

\item{method}{a method to perform singular-value decomposition when a dataset has missing values. See below for the explanation.}

\item{center.dat}{a logical specifying to center the input and denoised data. By default, \code{TRUE}.}

\item{scale.dat}{a logical specifying to scale the input and denoised data. By default, \code{FALSE}.}

\item{verbose}{a logical specifying to print the computational progress. By default, \code{FALSE}.}

\item{seed}{a seed for the random number generator.}

\item{...}{optional arguments.}
}
\value{
\code{preprocess_pca} returns a matrix of imputed and/or denoised data.
}
\description{
The data is denoised and the missing values are imputed by using the top r eigenmatrices.
After apply SVD on the centered and/or scaled data, the top r eigenmatrices are constructed where is r < min(n,m).
}
\details{
To impute missing values with PCA/SVD, two approximation methods are provided.
For method="nipals", a Non-linear Iterative Partial Least Squares (NIPALS) algorithm is used from \code{nipals} in the mixOmics package.
For method="em", a low-rank SVD approximation by the EM algorithm is used from \code{imputed.svd} in the bcv package.
}
\examples{
\dontrun{
data(cys_optm)
meta <- cys_optm[,1:4]
optm <- log(cys_optm[meta$Select,5:10])
optm <- t(scale(t(optm), scale=TRUE, center=TRUE))
days <- as.numeric(colnames(optm))

preprocessed_optm <- preprocess_pca(optm, timepoints = days, r=3, method=c("nipals","em"))
}
}
\references{
Troyanskaya, O., Cantor, M., Sherlock, G., Brown, P., Hastie, T., Tibshirani, R., Botstein, D. and Altman, R.B. (2001). Missing value estimation methods for DNA microarrays. Bioinformatics 17(6), 520–525.
}
\author{
Neo Christopher Chung \email{nchchung@gmail.com}
}
