% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise_spline.R
\name{preprocess_spline}
\alias{preprocess_spline}
\title{Preprocess the input data by fitting cubic splines}
\usage{
preprocess_spline(
  dat,
  timepoints = NULL,
  dof = c("cv", "cv.global"),
  center.dat = TRUE,
  scale.dat = FALSE,
  verbose = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{dat}{a time-series data matrix with \code{m} biomarkers as rows, over \code{n} time points (columns).}

\item{timepoints}{a vector of time points for columns of dat.}

\item{dof}{the degree of freedom in spline.smooth. By default, \code{cv} performs cross-validation per-variable.}

\item{center.dat}{a logical specifying to center the input and denoised data. By default, \code{TRUE}.}

\item{scale.dat}{a logical specifying to scale the input and denoised data. By default, \code{FALSE}.}

\item{verbose}{a logical specifying to print the computational progress. By default, \code{FALSE}.}

\item{seed}{a seed for the random number generator.}

\item{...}{optional arguments.}
}
\value{
\code{preprocess_spline} returns a matrix of imputed and/or denoised data.
}
\description{
The data is denoised and/or imputed by fitting cubic splines.
If a degree of freedom (\code{dof}) is set to \code{cv}, cross validation is performed on each variable to identify the optimal dof for that variable.
If \code{dof} set to \code{cv.global}, the mean of all cross-validated DoFs is used for all variables.
Lastly, \code{dof} can be set to a fixed numeric value predetermied by the user.
}
\examples{
\dontrun{
data(cys_optm)
meta <- cys_optm[,1:4]
optm <- log(cys_optm[meta$Select,5:10])
optm <- t(scale(t(optm), scale=TRUE, center=TRUE))
days <- as.numeric(colnames(optm))

preprocessed_optm <- preprocess_spline(optm, timepoints = days, dof="cv")
}
}
\author{
Neo Christopher Chung \email{nchchung@gmail.com}
}
