% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_dimred.R
\name{svd.vis}
\alias{svd.vis}
\title{Visualize two PCs}
\usage{
svd.vis(
  svd.dat,
  ident,
  dim.1 = 1,
  dim.2 = 2,
  pt.size = 1,
  pip = 1,
  no.axes = FALSE,
  coord_fixed = FALSE,
  title = ""
)
}
\arguments{
\item{svd.dat}{a SVD object. See corpcor::fastsvd.}

\item{ident}{identities of rows (e.g., clusters).}

\item{dim.1}{which dimension of PCA to display as the x-axis}

\item{dim.2}{which dimension of PCA to display as the y-axis}

\item{pt.size}{the size of data points}

\item{pip}{posterior inclusion probabilities}

\item{no.axes}{show axes. by default, FALSE}

\item{coord_fixed}{fix the coordinate. by default, FALSE}

\item{title}{provide a title}
}
\value{
\code{svd.vis} returns a ggplot2 object.
}
\description{
Plot two PCs where each rows (biomarkers) are colored by their group
}
\author{
Neo Christopher Chung \email{nchchung@gmail.com}
}
