% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_cluster.R
\name{vis_cluster}
\alias{vis_cluster}
\title{Visualize the grouped projectories}
\usage{
vis_cluster(
  dat,
  group = NULL,
  verbose = TRUE,
  seed = NULL,
  center.dat = TRUE,
  scale.dat = FALSE,
  ...
)
}
\arguments{
\item{dat}{a time-series data matrix with \code{m} biomarkers as rows, over \code{n} time points (columns).}

\item{group}{a vector defining the groups.}

\item{verbose}{a logical specifying to print the computational progress. By default, \code{FALSE}.}

\item{seed}{a seed for the random number generator.}

\item{center.dat}{a logical specifying to center the input and denoised data. By default, \code{TRUE}.}

\item{scale.dat}{a logical specifying to scale the input and denoised data. By default, \code{FALSE}.}

\item{...}{optional arguments.}
}
\value{
\code{vis_cluster} returns a ggplot2 object.
}
\description{
Plot the time course projectories that are grouped by clustering or user-defined groups.
}
\author{
Neo Christopher Chung \email{nchchung@gmail.com}
}
