% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_cluster.R
\name{vis_cluster_multi}
\alias{vis_cluster_multi}
\title{Visualize the projectories of two datasts}
\usage{
vis_cluster_multi(
  x,
  x.cluster,
  y,
  y.cluster,
  verbose = TRUE,
  print.plot = FALSE,
  return.separate = FALSE,
  ...
)
}
\arguments{
\item{x}{a first matrix with \code{m} biomarkers as rows, over \code{n} time points (columns).}

\item{x.cluster}{a clustering object from the first data matrix.}

\item{y}{a second matrix with \code{m} biomarkers as rows, over \code{n} time points (columns).}

\item{y.cluster}{a clustering object from the second data matrix.}

\item{verbose}{a logical specifying to print the computational progress. By default, \code{FALSE}.}

\item{print.plot}{a logical specifying to print the resulting plot onto the screen. By default, \code{FALSE}.}

\item{return.separate}{a logical specifying to return two ggplot2 objects separately. By default, \code{FALSE}.}

\item{...}{optional arguments.}
}
\value{
\code{vis_cluster_multi} returns a ggplot2 object.
}
\description{
Plot two sets of projectories next to each other, grouped by clusters.
}
\author{
Neo Christopher Chung \email{nchchung@gmail.com}
}
